corrected test, id column should be long not int

+++ src/test/scala-2.10/ch/unibas/dmi/dbis/adam/entity/EntityTestSuite.scala
@@ -96,7 +96,7 @@ class EntityTestSuite extends AdamBaseTest {
       }
 
       val dbNames = lb.toList.toMap
-      val templateNames = fieldTemplate.map(ft => (ft._1, ft._3)).toMap + (FieldNames.idColumnName -> "integer")
+      val templateNames = fieldTemplate.map(ft => (ft._1, ft._3)).toMap + (FieldNames.idColumnName -> "bigint")
 
       assert(dbNames.size == templateNames.size)
